Refuerza contrato público de usar y separa suite legacy
